refactor(cli): extract shared install helpers from the plugin command
Pulls the source-agnostic pieces (home path expansion, slug checks, git
subprocess runner, https single-file URL normalization with GitHub raw
rewriting, and the size-capped remote download) into install-utils so the
upcoming cline agent command reuses them instead of copying.
